third_party/blink/web_tests/fast/dom/Document/document-title-get-expected.txt

This test checks the implementation of getting the document.title attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test with no title set
PASS document.title is ""
Test with empty title
PASS document.title is ""
Test with only whitespace
PASS document.title is ""
Test with no whitespace
PASS document.title is "nowhitespacetitle"
Test with whitespace
PASS document.title is "one\u000b space"
Test with various whitespace lengths and fields
PASS document.title is "lots of whitespace and newlines"
Test with various length strings
PASS successfullyParsed is true

TEST COMPLETE
